+++ packages/super-editor/src/core/commands/insertTabNode.js
@@ -1,5 +1,5 @@
 export const insertTabNode = () => ({ tr, state }) => {
-    let newPos = tr.mapping.map(tr.selection.from);
+    let newPos = tr.selection.from;
     const node = state.schema.nodes.tab.create();
     tr.insert(newPos, node);
