--- packages/super-editor/src/extensions/tab-plugin/index.js
@@ -1,1 +1,0 @@
-export * from './tab-plugin';
